requirements.txt
Django==4.2.10
djangorestframework==3.14.0
drf-yasg==1.21.7
python-dotenv==1.0.0